refactor : layouts -> layout
