+++ tsconfig.json
@@ -16,7 +16,7 @@
     "paths": {
       "@/components/*": ["src/components/*"],
       "@/hooks/*": ["src/hooks/*"],
-      "@/layouts/*": ["src/layouts/*"],
+      "@/layout/*": ["src/layout/*"],
       "@/store/*": ["src/store/*"],
       "@/pages/*": ["src/pages/*"],
       "@/containers/*": ["src/layouts/*"],
@@ -26,6 +26,11 @@
       "@/public/*": ["public/*"],
     },
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/components/button/register"],
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    "src/components/button/register",
+  ],
   "exclude": ["node_modules"],
 }
